Add files via upload
The Employee Enquiry System is a web-based application designed to manage employee records efficiently. The system provides a platform for storing, retrieving, and updating essential employee details such as name, qualification, gender, and age. This project uses HTML and CSS for the front-end design, and Django, a robust Python-based framework, for the back-end, enabling seamless interaction with a relational database.